// EAM_CLI/src/app/models/tran/noti.model.ts
export class NotiModel {
  arbpl: string | null = null;
  qmnum: string = 'A';
  tplnr: string | null = null;
  eqart: string | null = null;
  equnr: string | null = null;
  priok: string | null = null;
  qmtxt: string | null = null;
  qmdetail: string | null = null;
  qmart: string = 'N2';
  iwerk: string | null = null;
  qmnam: string  = '';
  ingrp: string | null = null;
  staffSc: string | null = null;
  htbtbd: Date | null = null;
  lvtsd: Date | null = null;
  ltrmn: Date | null = null;
  qmdat: Date | null = null;
  isActive: boolean = true;
  htNbb: string | null = null;
  htDvql: string | null = null;
  htDvqlCd: string | null = null;
  htDvsd: string | null = null;
  htDvsdCd: string | null = null;
  htDvth: string | null = null;
  htDvthCd: string | null = null;
  htNdkt: string | null = null;
  htNddx: string | null = null;
  ntNbb: string | null = null;
  ntDvql: string | null = null;
  ntDvqlDes: string | null = null;
  ntDvqlCd: string | null = null;
  ntDvsd: string | null = null;
  ntDvsdDes: string | null = null;
  ntDvsdCd: string | null = null;
  ntDvth: string | null = null;
  ntDvthDes: string | null = null;
  ntDvthCd: string | null = null;
}